refactor(discord): split markdown formatter
